version avec un peu d'asservissement

+++ ausbee/include/AUSBEE/l298_driver.h
@@ -39,6 +39,6 @@ struct ausbee_l298_chip{
 enum AUSBEE_L298_DRIVER_ERROR ausbee_l298_init_chip(struct ausbee_l298_chip chip);
 void ausbee_l298_enable_chip(struct ausbee_l298_chip chip, uint8_t enable);
 void ausbee_l298_invert_output(struct ausbee_l298_chip chip, uint8_t enable);
-void ausbee_l298_set_duty_cycle(struct ausbee_l298_chip chip, uint8_t duty_cycle);
+void ausbee_l298_set_duty_cycle(struct ausbee_l298_chip chip, uint32_t duty_cycle);
 
 #endif
